Create README.md
This is some PHP code that I wrote many years ago for collective 2 API. I will admit it is not the best code but it worked just fine. I was never happy with Collective 2 and hence moved onto Interactive Brokers. I do recall that Collective 2 API was relatively easy to use in comparison to other APIs. 

I might be able to answer a question or two but I don’t remember most of it, without digging in. Hopefully this might help someone someday or not?